Removed DrawCost.  Modernized the Obj-C a bit.  Cleaned up a bit of documentation.  Exposed simultaneous loads in a few places.  SimpleIdentity is now 64 bit.
Former-commit-id: 1462d6992c9a8a8869274293a6ce459d8bc2a09b

+++ WhirlyGlobeSrc/WhirlyGlobeLib/include/Identifiable.h
@@ -31,7 +31,7 @@ namespace WhirlyKit
     an operation on a given ID rather than letting the developer muck around
     in the internals.
   */
-typedef unsigned long SimpleIdentity;
+typedef unsigned long long SimpleIdentity;
     
 /// This is the standard empty identity.  It means there were none of something
 ///  or it's just ignored.
